Рефакторинг фильтрации изображений
