+++ js/main.js
@@ -3,7 +3,7 @@ import {generateFullSizePost} from './full-size-mode.js';
 import {getData} from './api.js';
 import {showErrorMessageGet} from './message.js';
 import {setOnFormSubmit} from './form-loading.js';
-import {showFilterRicture} from './sort.js';
+import {showFilteredRictures} from './filters.js';
 import {debounce} from './util.js';
 
 const RERENDER_DELAY = 500;
@@ -12,7 +12,7 @@ getData()
   .then((pictures) => {
     generateThumbnails(pictures);
     generateFullSizePost(pictures);
-    showFilterRicture(pictures, debounce(generateThumbnails, RERENDER_DELAY));
+    showFilteredRictures(pictures, debounce(generateThumbnails, RERENDER_DELAY));
   })
   .catch(() => {
     showErrorMessageGet();
